Feature/redux_refactor #39 #40
 - Vertical, Feature Based Architecture  #39

 - Container and Presentational Component separation #40

 - Feature based architecture allows all relevant code to live within modularized folders (Containers Components, Reducers and Actions)
