+++ package.json
@@ -42,9 +42,15 @@
     "eslint-plugin-jsx-a11y": "^6.1.2",
     "eslint-plugin-prettier": "^3.0.0",
     "eslint-plugin-react": "^7.11.1",
+    "node-sass": "^4.10.0",
     "prettier": "1.14.2",
     "prettier-eslint": "^8.8.2"
   },
   "homepage": "./",
-  "browserslist": [">0.2%", "not dead", "not ie <= 11", "not op_mini all"]
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
