+++ src/App.js
@@ -4,10 +4,10 @@ import { Navbar, Jumbotron, Button } from 'react-bootstrap';
 
 // Redux
 import { Provider } from 'react-redux';
-import store from './store';
+import store from './state/store';
 
 // Components
-import MainContainer from './components/main/MainContainer';
+import HomePage from './features/homepage';
 
 import './App.css';
 
@@ -17,7 +17,7 @@ class App extends Component {
       <Provider store={store}>
         <Router>
           <div className="App">
-            <MainContainer />
+            <HomePage />
           </div>
         </Router>
       </Provider>
