--- src/actions/types.js
@@ -1,22 +1,0 @@
-// Whisper
-export const GET_WHISPER = 'GET_WHISPER';
-export const SEND_WHISPER_MESSAGE = 'SEND_WHISPER_MESSAGE';
-export const CREATE_LISTENER = 'CREATE_LISTENER';
-export const RECEIVED_MESSAGE = 'RECEIVED_MESSAGE';
-export const SET_WHISPER_PROVIDER = 'SET_WHISPER_PROVIDER';
-export const SET_WHISPER = 'SET_WHISPER';
-
-// Upload
-export const FILE_UPLOADED = 'FILE_UPLOADED';
-export const IPFS_ADD_FILE = 'IPFS_ADD_FILE';
-
-// Download
-export const FILE_DOWNLOADED = 'FILE_UPLOADED';
-export const IPFS_GET_FILE = 'IPFS_ADD_FILE';
-
-// Encryption
-export const GENERATE_ENCRYPTION_KEY = 'GENERATE_ENCRYPTION_KEY';
-export const ENCRYPT_FILE = 'ENCRYPT_FILE';
-
-// Decryption
-export const DECRYPT_FILE = 'DECRYPT_FILE';
