+++ src/features/download/index.js
@@ -1,0 +1,2 @@
+import DownloadContainer from './DownloadContainer';
+export default DownloadContainer;
