+++ src/features/homepage/index.js
@@ -1,0 +1,2 @@
+import HomepageContainer from './HomepageContainer';
+export default HomepageContainer;
