+++ src/features/providers/reducer.js
@@ -1,0 +1,20 @@
+import { SET_WHISPER_PROVIDER } from '../../state/types';
+// import isEmpty from '../utils/is-empty';
+
+const initialState = {
+  whisper: '',
+  ipfs: '',
+  web3: '',
+};
+
+export default function(state = initialState, action) {
+  switch (action.type) {
+    case SET_WHISPER_PROVIDER:
+      return {
+        ...state,
+        whisper: action.payload,
+      };
+    default:
+      return state;
+  }
+}
