+++ src/features/upload/index.js
@@ -1,0 +1,2 @@
+import UploadCardContainer from './UploadCardContainer';
+export default UploadCardContainer;
