+++ src/features/whisper/index.js
@@ -1,0 +1,2 @@
+import WhisperContainer from './WhisperContainer';
+export default WhisperContainer;
