+++ src/state/store.js
@@ -1,0 +1,19 @@
+import { createStore, applyMiddleware } from 'redux';
+import { composeWithDevTools } from 'redux-devtools-extension/developmentOnly';
+
+import thunk from 'redux-thunk';
+import rootReducer from './rootReducer';
+
+const initialState = {};
+const middleware = [thunk];
+
+const store = createStore(
+  rootReducer,
+  initialState,
+  composeWithDevTools(
+    applyMiddleware(...middleware),
+    //  window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+  ),
+);
+
+export default store;
